Change version to 5.2.0 Requires at least Refinerycms 4.1 Depends on Rails ~> 6.1
Use 'action_label' helper to for actions
Tests updated to reflect new controls
Use same testing setup as Refinery - including falcon
Add script to initialize tabs when editing a testimonial.
Simplify the select statement when selecting testimonials for a page
Split/Simplify the page decorators
Remove the 'distinct' condition for hmt association to pages.
This was causing a PG error, and it seems that if multiple occurences of an appearance/page tuple occurs it could be more easily fixed by deleting one of the occurrences.
(Some whitespace differences)